[update] skillsからfield/knowledgeを分離
